+++ xenon-common/src/main/java/com/vmware/xenon/common/http/netty/NettyChannelPool.java
@@ -544,7 +544,7 @@ public class NettyChannelPool {
         if (badContext == null) {
             return;
         }
-        Logger.getAnonymousLogger().info(
+        Logger.getAnonymousLogger().fine(
                 "replacing channel in bad state: " + badContext.toString());
         returnOrClose(badContext, true);
     }